Move GitHub recipes in its own subpackage
