+++ src/fastgithub/recipes/__init__.py
@@ -1,3 +1,3 @@
-"""A package that handles hook operation."""
+"""A package that handles hook operations."""
 
 from ._base import GithubRecipe, Recipe
