+++ src/fastgithub/recipes/github/__init__.py
@@ -1,0 +1,4 @@
+"""A package that handles GitHub hook operations."""
+
+from .autocreate_pr import AutoCreatePullRequest
+from .labels_from_commits import LabelsFromCommits
