+++ src/fastgithub/recipes/github/autocreate_pr.py
@@ -1,9 +1,8 @@
 from github.GithubException import GithubException
 
 from fastgithub.helpers.github import GithubHelper
+from fastgithub.recipes._base import GithubRecipe
 from fastgithub.types import Payload
-
-from ._base import GithubRecipe
 
 
 class AutoCreatePullRequest(GithubRecipe):
@@ -31,24 +30,3 @@ class AutoCreatePullRequest(GithubRecipe):
             except GithubException as ex:
                 if ex.status != 422:
                     raise ex
-
-
-LABEL_CONFIG: dict[str, list[str]] = {
-    "#nodraft": ["nodraft"],
-    "#fast": ["nodraft", "automerge", "autoapprove"],
-    "#release": ["nodraft", "automerge", "autorelease"],
-    "#furious": ["nodraft", "automerge", "autoapprove", "autorelease"],
-}
-
-
-class LabelFromCommit(GithubRecipe):
-    def __call__(
-        self,
-        payload: Payload,
-        labels_config: dict[str, list[str]] = LABEL_CONFIG,
-    ):
-        gh = GithubHelper(self.github, payload["repository"]["full_name"])
-        if not gh.rate_status.too_low():
-            pr = gh.repo.get_pull(payload["number"])
-            if labels := gh.extract_labels_from_pr(pr, labels_config):
-                gh.add_labels_to_pr(pr, labels)
